Extract unit tests from KeyPairUtilIT
